feat(frontend): add HomeContent component (#3556)
